requirements.txt
django
djangorestframework
Pillow
python-decouple
psycopg2-binary
drf-yasg

# для фильтрации и поиска
django-filter